2.0.0.2: - Reduced memory usage git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@189 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/frame/gpu_frame.h
@@ -32,7 +32,7 @@ public:
 	virtual std::vector<short>& audio_data();
 
 	virtual double alpha() const;
-	virtual void alpha(float value);
+	virtual void alpha(double value);
 
 	virtual double x() const;
 	virtual double y() const;
